Modification des liens de modification et de suppression pour uniquement l'utilisateur connecté

+++ view/utilisateur/detail.php
@@ -16,12 +16,20 @@
       Prénom: <span class="text-info"><?php echo $uPrenom;?></span>
     </div>
   </div>
-  <div class="row">
-    <div class="col-sm-offset-1 col-sm-4">
-      <a href="?action=update&controller=utilisateur&login=<?=$uLoginURL ?>" class="btn btn-warning btn-block" role="button">Modifier</a>
-    </div>
-    <div class="col-sm-offset-2 col-sm-4">
-        <a href="?action=delete&controller=utilisateur&login=<?=$uLoginURL ?>" class="btn btn-danger btn-block" role="button">Supprimer</a>
-    </div>
-  </div>
+    <?php
+
+    if(Session::is_user($uLogin)){
+        ?>
+        <div class="row">
+            <div class="col-sm-offset-1 col-sm-4">
+                <a href="?action=update&controller=utilisateur&login=<?=$uLoginURL ?>" class="btn btn-warning btn-block" role="button">Modifier</a>
+            </div>
+            <div class="col-sm-offset-2 col-sm-4">
+                <a href="?action=delete&controller=utilisateur&login=<?=$uLoginURL ?>" class="btn btn-danger btn-block" role="button">Supprimer</a>
+            </div>
+        </div>
+    <?php
+    }
+
+    ?>
 </div>
